+++ example/src/ListScreen.tsx
@@ -1,14 +1,14 @@
 import React from 'react';
-import { ScrollView, Text, View } from 'react-native';
-import { Avatar, Button, Card } from 'react-native-paper';
-import { SafeAreaView, useSafeArea } from 'react-native-safe-area-context';
+import {ScrollView, Text, View} from 'react-native';
+import {Avatar, Button, Card} from 'react-native-paper';
+import {useSafeAreaInsets} from 'react-native-safe-area-context';
 
 const LeftContent = (props: any) => <Avatar.Icon {...props} icon="folder" />;
 
 export function SettingsScreen() {
-  const {top} = useSafeArea();
+  const {bottom} = useSafeAreaInsets();
   return (
-    <View style={{flex: 1, marginTop: top}}>
+    <View style={{flex: 1, marginBottom: bottom}}>
       <ScrollView>
         <View>
           <Card elevation={5} style={{margin: 5}}>
@@ -60,7 +60,8 @@ export function SettingsScreen() {
             </Card.Actions>
           </Card>
         </View>
-        {/* <View style={{height: 120}} /> -> Set bottom view to allow scrolling to top if you set bottom-bar position absolute */}
+        {/* -> Set bottom view to allow scrolling to top if you set bottom-bar position absolute */}
+        <View style={{height: 90}} />
       </ScrollView>
     </View>
   );