+++ src/__test__/tab.bar.button.spec.tsx
@@ -7,9 +7,16 @@ import { BarButton, TabBarButton } from '../components/tab.bar.button';
 const tabBarIcon = (props: any) => <View {...props} />;
 
 describe('TabBarButton', () => {
+  beforeAll(() => {
+    jest.useFakeTimers();
+  });
+  afterAll(() => {
+    jest.useRealTimers();
+  });
   it("Should render well when it's Focus", () => {
     const tree = renderer.create(
       <BarButton
+        mode="default"
         index={0}
         isFocused
         options={{ tabBarIcon }}
@@ -17,11 +24,13 @@ describe('TabBarButton', () => {
         onLongPress={jest.fn()}
       />
     );
+    jest.runAllTimers();
     expect(tree.toJSON()).toMatchSnapshot();
   });
   it('Should render well it is not focused', () => {
     const tree = renderer.create(
       <BarButton
+        mode="default"
         index={0}
         isFocused={false}
         options={{ tabBarIcon }}
@@ -29,11 +38,13 @@ describe('TabBarButton', () => {
         onLongPress={jest.fn()}
       />
     );
+    jest.runAllTimers();
     expect(tree.toJSON()).toMatchSnapshot();
   });
   it("Should render well when it's Focus", () => {
     const tree = renderer.create(
       <TabBarButton
+        mode="default"
         index={0}
         isFocused
         options={{ tabBarIcon }}
@@ -41,11 +52,13 @@ describe('TabBarButton', () => {
         onLongPress={jest.fn()}
       />
     );
+    jest.runAllTimers();
     expect(tree.toJSON()).toMatchSnapshot();
   });
   it('Should render well it is not focused', () => {
     const tree = renderer.create(
       <TabBarButton
+        mode="default"
         index={0}
         isFocused={false}
         options={{ tabBarIcon }}
@@ -53,6 +66,21 @@ describe('TabBarButton', () => {
         onLongPress={jest.fn()}
       />
     );
+    jest.runAllTimers();
+    expect(tree.toJSON()).toMatchSnapshot();
+  });
+  it('Should render well it is not focused', () => {
+    const tree = renderer.create(
+      <TabBarButton
+        mode="square"
+        index={0}
+        isFocused={false}
+        options={{ tabBarIcon }}
+        onPress={jest.fn()}
+        onLongPress={jest.fn()}
+      />
+    );
+    jest.runAllTimers();
     expect(tree.toJSON()).toMatchSnapshot();
   });
 });