+++ src/components/tab.bar.button.tsx
@@ -1,13 +1,21 @@
 import React, { memo, useEffect, useState } from 'react';
-import { Animated, TouchableOpacity, View } from 'react-native';
+import {
+  Animated,
+  StyleProp,
+  Text,
+  TouchableOpacity,
+  View,
+} from 'react-native';
 
 import { BottomTabNavigationOptions } from '@react-navigation/bottom-tabs';
 
 import { style } from '../styles/tab.bar.button.styles';
 
 const AnimatedTouchable = Animated.createAnimatedComponent(TouchableOpacity);
+const AnimatedText = Animated.createAnimatedComponent(Text);
 
 interface Props {
+  mode: 'default' | 'square';
   index: number;
   isFocused: boolean;
   onPress: () => void;
@@ -16,6 +24,7 @@ interface Props {
   inactiveTintColor?: string;
   activeTintColor?: string;
   springConfig?: Animated.SpringAnimationConfig;
+  focusedButtonStyle?: StyleProp<any>;
 }
 
 export const defaultSpringConfig = {
@@ -66,7 +75,7 @@ export const BarButton: React.FC<Props> = memo(
           ]}
           onLongPress={onLongPress}
         >
-          <View style={{ zIndex: 12 }}>
+          <View style={{ zIndex: 12, alignItems: 'center' }}>
             {options.tabBarIcon && !isFocused ? (
               options.tabBarIcon({
                 focused: isFocused,
@@ -75,6 +84,24 @@ export const BarButton: React.FC<Props> = memo(
               })
             ) : (
               <View />
+            )}
+            {options.tabBarLabel && (
+              <AnimatedText
+                style={[
+                  {
+                    marginTop: 2,
+                    color: inactiveTintColor,
+                  },
+                  {
+                    opacity: animationValueThreshold.interpolate({
+                      inputRange: [0.5, 1],
+                      outputRange: [0, 1],
+                    }),
+                  },
+                ]}
+              >
+                {options.tabBarLabel}
+              </AnimatedText>
             )}
           </View>
         </AnimatedTouchable>
@@ -91,6 +118,8 @@ export const TabBarButton: React.FC<Props> = memo(
     onLongPress,
     activeTintColor,
     springConfig,
+    focusedButtonStyle,
+    mode,
   }) => {
     const [animationValueThreshold] = useState(new Animated.Value(0));
 
@@ -106,12 +135,14 @@ export const TabBarButton: React.FC<Props> = memo(
       <View style={style.wrapper}>
         <AnimatedTouchable
           accessibilityRole="button"
+          accessibilityComponentType="Button"
           accessibilityLabel={options.tabBarAccessibilityLabel}
           testID={options.tabBarTestID}
           onPress={onPress}
           style={[
             {
               ...style.focusedButton,
+              ...(mode === 'square' ? style.squareFocusedButton : {}),
               backgroundColor: activeTintColor || 'white',
               transform: [
                 {
@@ -122,6 +153,7 @@ export const TabBarButton: React.FC<Props> = memo(
                 },
               ],
             },
+            isFocused ? focusedButtonStyle : {},
           ]}
           onLongPress={onLongPress}
         >